# Speed up csv_export for large datasets (by throttling yield)