## Summary

* Speeds up large CSV exports by throttling `yield` calls to once every 1000 rows instead of every row
* Manual verification with 100,000 row exports on both high-end and low-end hardware
* No UI changes

**Benchmark Results**

**Mac Mini M2 Pro:**

- **Before:** 100k rows in 78.7s  
- **After:**  100k rows in 24.4s  
- 69% (~3.2x) faster

**RACHEL-Plus (1.1GHz Celeron):**

- **Before:** 100k rows in 1512s  
- **After:**  100k rows in 193s  
- 87% (~7.8x) faster

Timing was captured by inserting `time.time()`-based print statements in `csv_export.py` No external profiling tools were used.

## References

* No related issue (optimized after running into trouble in the field)

## Reviewer guidance

* Add ContentSessionLog data if your db has < 100k (see: devscripts/generate_fake_sessions.py)
* Under Facility > Data > Session logs, generate a session log 
* Confirm export completes successfully
* Note improved export time and fewer updates to the job_storage db
 

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
  - Added a script for generating large numbers of fake session logs to assist with testing and benchmarking in development environments.

- **Performance Improvements**
  - Improved the efficiency of CSV export by batching progress updates, resulting in faster exports for large datasets.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->