@needlestack 2025-05-07
In case anyone is wondering the yield still gets called every 0.7 seconds on the Mac and 2.0 seconds on the RACHEL, so updates to jobs_storage.db should still be sufficiently fresh