@rtibbles 2025-05-09
Hi @needlestack - thanks for this. This seems like a valuable performance improvement, so I am just doing some double checking in review so that we can get this into the next patch release of Kolibri.